electron-main: Immediatelly set the eventIndex variable to null when closing.
The react-sdk doesn't await the closing of the event index, this is done
so because of limitations in the lifecycle module. This wasn't a problem
since we used to just set the eventIndex variable to null. Nowadays we
wait for Tantivy to shutdown using a close() method on the index.

To avoid this being called multiple times while we're already closing the
index set the eventIndex variable to null before awaiting.

This fixes https://github.com/vector-im/riot-web/issues/12838.